// src/generated/models/Geek_incidentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Geek_incidentsgeek_priority = {
  100000000: '緊急',
  100000001: '高',
  100000002: '中',
  100000003: '低'
} as const;
export type Geek_incidentsgeek_priority = keyof typeof Geek_incidentsgeek_priority;
export const Geek_incidentsgeek_status = {
  100000000: '新規',
  100000001: '対応中',
  100000002: '保留',
  100000003: '解決済',
  100000004: 'クローズ'
} as const;
export type Geek_incidentsgeek_status = keyof typeof Geek_incidentsgeek_status;
export const Geek_incidentsstatecode = {
  0: 'アクティブ',
  1: '非アクティブ'
} as const;
export type Geek_incidentsstatecode = keyof typeof Geek_incidentsstatecode;
export const Geek_incidentsstatuscode = {
  1: 'アクティブ',
  2: '非アクティブ'
} as const;
export type Geek_incidentsstatuscode = keyof typeof Geek_incidentsstatuscode;

export interface Geek_incidentsBase {
  "geek_assignedtoid@odata.bind"?: string;
  geek_description?: string;
  geek_duedate?: string;
  "geek_incidentcategoryid@odata.bind"?: string;
  geek_incidentid: string;
  "geek_locationid@odata.bind"?: string;
  geek_name: string;
  geek_priority?: Geek_incidentsgeek_priority;
  geek_status?: Geek_incidentsgeek_status;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Geek_incidentsstatecode;
  statuscode?: Geek_incidentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Geek_incidents extends Geek_incidentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  geek_assignedtoidname?: string;
  geek_assignedtoidyominame?: string;
  geek_incidentcategoryidname?: string;
  geek_locationidname?: string;
  geek_priorityname?: string;
  geek_statusname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  geek_assignedtoid?: object;
  _geek_assignedtoid_value?: string;
  geek_incidentcategoryid?: object;
  _geek_incidentcategoryid_value?: string;
  geek_locationid?: object;
  _geek_locationid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}